Add error messages to data validation function
A new field `Message` and function `fmtMessage` have been added to the `Validator` struct. `fmtMessage` produces detailed error messages based on the field and tag causing validation to fail, which are then returned as part of the `ValidationErrorResponse`. User input is now more robust and error feedback is more detailed, enhancing the functionality of the data validation process.