Update validation errors handling and response
Refactored the Validator to store validation errors in a map instead of a struct and structured the errors based on the field and tag causing validation failure. Also changed the error response from `badRequestErrorResponse` to `failedValidationResponse` in the movie handler file providing a more relevant http status and a detailed error message to improve the robustness of the validation process.