Add PostgreSQL database connection and Docker support
Introduced support for PostgreSQL database with the necessary configuration being added to the API application. Implemented a function to open a connection to this database. The functionality to run the API server has been containerized via a Dockerfile and can be controlled using the provided Makefile. Added a new docker-compose file to manage multiple services, including a PostgreSQL database service.